+++ routes/reviews.js
@@ -7,18 +7,9 @@ const Review = require("../models/review");
 const ExpressError = require('../utils/ExpressError');
 const catchAsync = require('../utils/catchAsync');
 
+const { validateReview } = require("../middleware")
 
-const { reviewSchema } = require('../schemas.js')
 
-const validateReview = (req, res, next) => {
-    const { error } = reviewSchema.validate(req.body);
-    if (error) {
-        const msg = error.details.map(el => el.message).join(',')
-        throw new ExpressError(msg, 400)
-    } else {
-        next();
-    }
-}
 
 router.post("/", validateReview, catchAsync(async (req, res) => {
     const campground = await Campground.findById(req.params.id);
@@ -34,7 +25,7 @@ router.delete("/:reviewId", catchAsync(async (req, res) => {
     const { id, reviewId } = req.params;
     await Campground.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
     await Review.findByIdAndDelete(req.params.reviewId);
-    res.flash("success", "Successfully deleted a review!")
+    req.flash("success", "Successfully deleted a review!")
     res.redirect(`/campgrounds/${id}`);
 }))
